+++ src/dispatch.cc
@@ -1,6 +1,6 @@
 /*
 	dtee - run a program with standard output and standard error copied to files
-	Copyright 2018  Simon Arlott
+	Copyright 2018,2021  Simon Arlott
 
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU General Public License as published by
@@ -19,12 +19,13 @@
 
 #include <cstddef>
 #include <string>
+#include <vector>
 
 using ::std::string;
 
 namespace dtee {
 
-Dispatch::Dispatch(const std::list<std::shared_ptr<Output>> &outputs, const std::list<std::shared_ptr<ResultHandler>> &result_handlers)
+Dispatch::Dispatch(const std::vector<std::shared_ptr<Output>> &outputs, const std::vector<std::shared_ptr<ResultHandler>> &result_handlers)
 		: outputs_(outputs), result_handlers_(result_handlers) {
 }